requirements.txt
streamlit==1.31.0
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
matplotlib==3.8.0
plotly==5.18.0
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
surprise==0.1
nltk==3.8.1
scipy==1.11.3
wordcloud==1.9.2
beautifulsoup4==4.12.2
lxml==4.9.3 